server-yolo/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
onnxruntime==1.22.1
onnx<1.18
ultralytics==8.3.6
numpy==1.26.4
pillow==10.4.0
opencv-python-headless==4.10.0.84
python-multipart
# Added to force Docker rebuild
